course_100_days_of_code/day_32/email_downloader_and_parser/pyproject.toml
[tool.poetry]
name = "email-downloader-and-parser"
version = "0.1.0"
description = ""
authors = ["alessandro-maccario <a.maccario1@campus.unimib.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pandas = "^2.2.3"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
